crewai-masumi-quickstart-template/requirements.txt
fastapi
uvicorn
python-dotenv
crewai
masumi
pydantic
python-multipart
httpx
crewai-tools
minswap-py